Refactor: mv insert_data/* -> parser/ & clean up parser.h & insert_data.h
