--- src/manda/insert_data/Makefile
@@ -1,18 +1,0 @@
-ifndef TOPDIR
-	TOPDIR = $(abspath ../../../)
-endif
-ifndef SRCDIR
-	SRCDIR = $(abspath ../)
-endif
-
-NAME := insert_data.a
-HEAD := insert_data.h
-SRCS := insert_lr_table.c \
-		insert_grammar.c \
-		insert_util.c \
-
-LDLIBS := -lft
-
-include $(TOPDIR)/include_make/Variable.mk
-include $(TOPDIR)/include_make/Link.mk
-include $(TOPDIR)/include_make/Recipe_subsrc.mk
