+++ src/manda/parser/insert_lr_table.c
