+++ src/manda/parser/parser.h
@@ -6,7 +6,7 @@
 /*   By: taerakim <taerakim@student.42seoul.kr>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/03/22 14:01:48 by taerakim          #+#    #+#             */
-/*   Updated: 2024/03/28 14:29:43 by taerakim         ###   ########.fr       */
+/*   Updated: 2024/03/28 15:20:37 by taerakim         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,51 +14,39 @@
 # define PARSER_H
 
 # include <stdbool.h>
-/* ******************************** LITERAL ********************************* */
-# define S_PIPE "|"
-# define S_AND_IF "&&"
-# define S_OR_IF "||"
-# define S_LESS "<"
-# define S_DLESS "<<"
-# define S_GREAT ">"
-# define S_DGREAT ">>"
-# define S_LPAREN "("
-# define S_RPAREN ")"
 
 /* ********************************** TYPE ********************************** */
-# define TERMINAL 0
-# define NON_TERMINAL 1
+typedef enum e_kind
+{
+	none = -1,
+	terminal,
+	non_terminal
+}	t_kind;
 
-enum e_terminal
+typedef enum e_terminal
 {
 	and_if = 0,
 	or_if,
 	pipe,
 	lparen,
 	rparen,
-	// name,
 	word,
-	// assignment_word,
 	less,
 	great,
 	dgreat,
-	dless
-	// dollar_sign
-};
+	dless,
+	dollar_sign
+}	t_termi;
 
-enum e_non_terminal
+typedef enum e_non_terminal
 {
 	complete_command = 0,
 	and_or,
 	pipeline,
 	command,
-	// compound_command,
 	subshell,
-	// name,
 	simple_command,
-	// cmd_name,
 	cmd_word,
-	// cmd_prefix,
 	cmd_suffix,
 	redirect_list,
 	io_redirect,
@@ -66,15 +54,12 @@ enum e_non_terminal
 	filename,
 	io_here,
 	here_end
-};
-
-typedef enum e_terminal		t_termi;
-typedef enum e_non_terminal	t_ntermi;
+}	t_ntermi;
 
 /* ********************************* STRUCT ********************************* */
-# define FRONT 0
+# define LEFT 0
 # define MID 1
-# define BACK 2
+# define RIGTH 2
 
 typedef struct s_token
 {
@@ -88,5 +73,16 @@ typedef struct s_parse_tree
 	bool		child_type[3];
 	void		*child[3];
 }				t_parse_tree;
+
+/* ******************************** LITERAL ********************************* */
+# define S_PIPE "|"
+# define S_AND_IF "&&"
+# define S_OR_IF "||"
+# define S_LESS "<"
+# define S_DLESS "<<"
+# define S_GREAT ">"
+# define S_DGREAT ">>"
+# define S_LPAREN "("
+# define S_RPAREN ")"
 
 #endif
